Fix all shellchck warnings, start fixing man pages

+++ scripts/sdk/install.sh
@@ -19,17 +19,19 @@ sdk_installSdkPackages() {
   local -r sdkPkgEnvName="SDKVM_${sdk^^}_PACKAGES"
   if sdk_hasAction "$sdk" "installPackages"; then
     printInfo "Installing SDK global packages $sdk/$version"
-    cd "$targetDir"
+    cd "$targetDir" || error "Dir does not exist $targetDir"
     sdk_execute "$sdk" installPackages "$version" "$targetDir" "$sdkDir"
     printDebug "SDK global packages installed successffuly $sdk/$version"
   elif sdk_hasAction "$sdk" "installPackage" && [ -n "${!sdkPkgEnvName}" ]; then
     printInfo "Installing SDK global packages $sdk/$version"
-    cd "$targetDir"
+    cd "$targetDir" || error "Dir does not exist $targetDir"
     for pkg in ${!sdkPkgEnvName}; do
       if [ -n "$pkg" ]; then
-        sdk_execute "$sdk" installPackage "$pkg" "$version" "$targetDir" "$sdkDir" &&
-          printInfo "Package installed successfully: $pkg" ||
+        if sdk_execute "$sdk" installPackage "$pkg" "$version" "$targetDir" "$sdkDir"; then
+          printInfo "Package installed successfully: $pkg"
+        else
           printWarn "Could not install package: $pkg"
+        fi
       fi
     done
     printDebug "SDK global packages installed $sdk/$version"
@@ -44,7 +46,7 @@ sdk_uninstallSdkVersion() {
   if sdk_isLocalSdkVersion "$sdk" "$version"; then
     printInfo "Uninstalling SDK $sdk/$version"
     sdk_isEnabled "$sdk" "$version" && sdk_disable "$sdk" "$version"
-    rm -rf "$SDKVM_LOCAL_SDKS_DIR/$sdk/$version"
+    rm -rf "${SDKVM_LOCAL_SDKS_DIR:?}/$sdk/$version"
     printDebug "SDK uninstalled successffuly $sdk/$version"
   else
     printInfo "SDK is not installed $sdk/$version. Skipping..."
@@ -60,7 +62,7 @@ sdk_uninstallSdk() {
     for version in $versions; do
       sdk_uninstallSdkVersion "$sdk" "$version"
     done
-    rm -rf "$SDKVM_LOCAL_SDKS_DIR/$sdk"
+    rm -rf "${SDKVM_LOCAL_SDKS_DIR:?}/$sdk"
   else
     printInfo "SDK is not installed $sdk. Skipping..."
   fi
